Clean up features, unneeded drivers
