+++ src/lib.rs
@@ -3,29 +3,29 @@
 #![cfg_attr(target_os="none", no_std)]
 #![no_main]
 
-#[cfg(not(feature = "std"))]
+#[cfg(target_os = "none")]
 extern crate alloc;
 
 pub mod telemetry;
 pub mod settings;
 pub mod state_estimation;
 
-#[cfg(feature = "no_std")]
+#[cfg(target_os = "none")]
 use defmt_rtt as _; // global logger (TODO)
 
-#[cfg(feature = "no_std")]
+#[cfg(target_os = "none")]
 use panic_probe as _;
 
 /// same panicking *behavior* as `panic-probe` but doesn't print a panic message
 /// this prevents the panic message being printed *twice* when `defmt::panic` is invoked
 #[defmt::panic_handler]
-#[cfg(feature = "no_std")]
+#[cfg(target_os = "none")]
 fn panic() -> ! {
     cortex_m::asm::udf()
 }
 
 /// Terminates the application and makes `probe-run` exit with exit-code = 0
-#[cfg(feature = "no_std")]
+#[cfg(target_os = "none")]
 pub fn exit() -> ! {
     loop {
         cortex_m::asm::bkpt();
