+++ src/telemetry.rs
@@ -1,25 +1,25 @@
 //! Contains data structures that the flight computer shares with the outside world,
 //! namely the ground station software, as well as common (de)serialization code.
 
-#[cfg(not(feature = "std"))]
+#[cfg(target_os = "none")]
 use alloc::string::{String, ToString};
-#[cfg(not(feature = "std"))]
+#[cfg(target_os = "none")]
 use alloc::vec::Vec;
+
+#[cfg(not(target_os = "none"))]
+use std::string::{String, ToString};
+#[cfg(not(target_os = "none"))]
+use std::vec::Vec;
 
 #[cfg(target_os="none")]
 use core::f32::consts::PI;
-#[cfg(not(target_os="none"))]
+#[cfg(not(target_os = "none"))]
 use std::f32::consts::PI;
 
 #[cfg(target_os = "none")]
 use core::hash::Hasher;
 #[cfg(not(target_os = "none"))]
 use std::hash::Hasher;
-
-#[cfg(feature = "std")]
-use std::string::{String, ToString};
-#[cfg(feature = "std")]
-use std::vec::Vec;
 
 use nalgebra::*;
 use serde::{Deserialize, Serialize};
